+++ src/rsfml/window/window.rs
@@ -238,7 +238,7 @@ impl Window {
                 3   => event::GainedFocus,
                 4   => {
                     event::TextEntered{
-                        code : self.event.p1 as char
+                        code : (self.event.p1 as u8) as char
                     }
                 },
                 5   => {
